# swagger/ecomm.yaml
openapi: 3.0.3
info:
  title: Swagger ecomm - OpenAPI 3.0
  description: |-
    An api used for Alura level Up - Next Dev
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
servers:
  - url: http://localhost:3000
  - url: http://localhost:3001
...
paths:
...
  /api/admin/products:
    post:
      tags:
        - products
      summary: Creater a product
      description: Returns product created
      operationId: createProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/productDTO'
      responses:
        '200':
          description: successful operation
          content:
           application/json:
              schema:
                $ref: "#/components/schemas/product"
        '404':
          description: Nenhum produto encontrado
          content:
            application/json:
              example: "Not Found"
...
components:
  schemas:
...
    product:
      type: object
      properties:
        _id:
          type: string
        nome:
          type: string
        descricao:
          type: string
        slug:
          type: string
        precoUnitario:
          type: number
        quantidadeEmEstoque:
          type: integer
        categoria:
          type: object
          properties:
            id:
              type: string
    productDTO:
      type: object
      properties:
        nome:
          type: string
        descricao:
          type: string
        slug:
          type: string
        precoUnitario:
          type: number
        quantidadeEmEstoque:
          type: integer
        categoria:
          type: object
          properties:
            id:
              type: string
...